requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
bsedata==0.6.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
curl_cffi==0.13.0
dateutils==0.6.12
distro==1.9.0
ecdsa==0.19.1
fastapi==0.119.0
frozendict==2.4.6
frozenlist==1.8.0
google-auth==2.42.1
google-genai==1.47.0
greenlet==3.2.4
groq==0.33.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.7.0
multitasking==0.0.12
nsepy==0.8
nsepython==2.97
nsetools==2.0.1
numpy==2.3.3
pandas==2.3.3
passlib==1.7.4
peewee==3.18.2
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
upstox-python-sdk==2.18.0
urllib3==2.5.0
uuid==1.30
uvicorn==0.37.0
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
yfinance==0.2.66
